// backend/src/users/enums/project-team-role.enum.ts
export enum ProjectTeamRole {
  PROJECT_OWNER = 'project_owner',
  PROJECT_CREATOR = 'project_creator',
  PROJECT_MANAGER = 'project_manager',
  DESIGNER = 'designer',
  DEVELOPER = 'developer',
  FRONTEND_DEVELOPER = 'frontend_developer',
  BACKEND_DEVELOPER = 'backend_developer',
  QA = 'qa',
  MARKETER = 'marketer',
  BUSINESS_ANALYST = 'business_analyst',
}